Description editor and double-click behaviour fixes
Former-commit-id: 8eb548864256165900573f3f492e8195c871645c [formerly c97c25989704b3f54f4ce8c4fafd0dd7c5083488] [formerly 178017f334ec06d06d0110d69ef9876c392ea1b7 [formerly dd7420f03e263a24759591fbea3f522bdd3a17d4]]
Former-commit-id: f6d1ffde3e245d09f05dee610da8907d26bc914c [formerly 24aa0f196347fb4a652e936fdc07c04e8f6c422b]
Former-commit-id: 47ef3b7be4dd6d88d808e61f95e944d1a5a99aa7

+++ imagecache.py
@@ -300,7 +300,7 @@ class ImageCache(QtCore.QObject):
                 sum(g) / float(len(g)),
                 sum(b) / float(len(b))
             )
-        average_color.setAlpha(average_color.alpha() / 2.0)
+        average_color.setAlpha(int(average_color.alpha() / 2))
         return average_color
 
     def generate_thumbnails(self, indexes, overwrite=False):
